jwt-server/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==3.2.2
blinker==1.9.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.121.0
flask==3.1.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.11.0
pyjwt==2.10.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
werkzeug==3.1.3
